[ACS-2557] Moved test methods to more appropriate test class
